backend/apps/apis/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import APIViewSet

router = DefaultRouter()
router.register(r'apis', APIViewSet, basename='api')

urlpatterns = [
    path('', include(router.urls)),
]



