backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-filter==23.3
Pillow==10.1.0
httpx==0.25.2
django-crontab==0.7.1
pymysql==1.1.0
cryptography==41.0.7
APScheduler==3.10.4

